requirement.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
distro==1.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.26.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kagglehub==0.3.1
keras==3.6.0
keras-nlp==0.15.1
kiwisolver==1.4.5
langcodes==3.3.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.2
openai==1.43.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.9.11
requests==2.31.0
rich==13.9.1
safetensors==0.4.5
scipy==1.13.0
seaborn==0.13.2
selenium==4.21.0
setuptools==69.0.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
thinc==8.2.1
tiktoken==0.7.0
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.45.2
trio==0.25.1
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webdriver-manager==4.0.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
